if gid is already existing on a specified location, it is ignored. = update amount NOW
Inventory Management System
- Changed Lot.source (Lot) to Lot.sourceId (Integer)
- Changed Transaction.quantity and Transaction.prevAmount from Integer to Double
- Refactored the code based on the above changes
- Updated InventoryService.addLots() to include the parameters amount (for quantity) and sourceId (listId)
- Moved IMS pojos to org.generationcp.middleware.pojos.ims
- Moved IMS DAOs to org.generationcp.middleware.dao.ims

Reviewer: Daniel Villafuerte

Issues: GCP-8820, GCP-8802

+++ src/main/java/org/generationcp/middleware/manager/api/InventoryDataManager.java
@@ -16,8 +16,8 @@ import java.util.Set;
 
 import org.generationcp.middleware.domain.inventory.InventoryDetails;
 import org.generationcp.middleware.exceptions.MiddlewareQueryException;
-import org.generationcp.middleware.pojos.Lot;
-import org.generationcp.middleware.pojos.Transaction;
+import org.generationcp.middleware.pojos.ims.Lot;
+import org.generationcp.middleware.pojos.ims.Transaction;
 import org.generationcp.middleware.pojos.report.LotReportRow;
 import org.generationcp.middleware.pojos.report.TransactionReportRow;
 
@@ -139,7 +139,7 @@ public interface InventoryDataManager{
      * @return The actual transaction balance of all the specified Lot
      * @throws MiddlewareQueryException the middleware query exception
      */
-    Long getActualLotBalance(Integer lotId) throws MiddlewareQueryException;
+    Double getActualLotBalance(Integer lotId) throws MiddlewareQueryException;
 
     /**
      * Returns the available transaction balance of the Lot with the specified
@@ -151,7 +151,7 @@ public interface InventoryDataManager{
      * @return The available transaction balance of all the specified Lot
      * @throws MiddlewareQueryException the middleware query exception
      */
-    Long getAvailableLotBalance(Integer lotId) throws MiddlewareQueryException;
+    Double getAvailableLotBalance(Integer lotId) throws MiddlewareQueryException;
 
     /**
      * Given a valid Lot object, add it as a new record to the database. It is
